fix(tradepro): StrategySignal.confidence is Int?, drop .toDouble()
compileDebugKotlin failed: BacktestLabViewModel:163 passed Double to the
Int? confidence param. setup.confidence is already Int, so pass it directly.

+++ app/src/main/java/com/foxtrader/app/feature/backtest/presentation/BacktestLabViewModel.kt
@@ -160,7 +160,7 @@ class BacktestLabViewModel @Inject constructor(
             entry = setup.entry,
             stopLoss = setup.stopLoss,
             takeProfit = setup.target2,
-            confidence = setup.confidence.toDouble(),
+            confidence = setup.confidence,
             setupType = BacktestStrategyTemplate.TRADEPRO.displayName,
         )
     }
